backend/requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
python-dotenv

# Core analytics libraries
pandas
statsmodels
scikit-learn
geopandas
aiohttp
s3fs
scipy
matplotlib
folium
h3

# Task Queue
celery
redis
